Issue-39: Rearchitect ConvivialProfiler Library.
