+++ package.json
@@ -23,6 +23,14 @@
     "webpack-cli": "^4.9.2",
     "webpack-dev-server": "^4.7.4"
   },
+  "engines": {
+    "npm": ">=6.14.0 <7.0.0",
+    "node": ">=12.22.0 <13.0.0"
+  },
+  "browserslist": [
+    "> 0.2%",
+    "not dead"
+  ],
   "dependencies": {},
   "repository": {
     "type": "git",
