--- src/profiler_destination/unset.js
@@ -1,24 +1,0 @@
-/**
- * @file
- * Convivial Profiler library destination plugin.
- *
- * Copyright Morpht Pty Ltd 2020-2022.
- */
-
-(function (window, localStorage) {
-  window.convivialProfiler = window.convivialProfiler || {};
-  window.convivialProfiler.profilerDestination = window.convivialProfiler.profilerDestination || {};
-  window.convivialProfiler.profilerDestination.unset = function (profiler, destination, values) {
-    var storage_value = window.convivialProfiler._getValue(destination.storage_key);
-    if (storage_value) {
-      // Unset the data in localstorage if its applicable.
-      if (destination.target_location.localstorage === 'localstorage') {
-        localStorage.setItem(storage_value, '0');
-      }
-      // Unset the data in cookie if its applicable.
-      if (destination.target_location.cookie === 'cookie') {
-        window.convivialProfiler._setCookie(storage_value, '0');
-      }
-    }
-  }
-})(window, localStorage);
