--- src/profiler_processor/temp.js
@@ -1,17 +1,0 @@
-/**
- * @file
- * Convivial Profiler library processor plugin.
- *
- * Copyright Morpht Pty Ltd 2020-2022.
- */
-
-(function (window) {
-  window.convivialProfiler = window.convivialProfiler || {};
-  window.convivialProfiler.profilerProcessor = window.convivialProfiler.profilerProcessor || {};
-  window.convivialProfiler.profilerProcessor.temp = function (profiler, processor, values) {
-    // Store values temporarily.
-    values.forEach(value => {
-      window.convivialProfiler._setValue('temp', processor.storage_key, {value: value});
-    });
-  }
-})(window);
