--- src/profiler_source/meta.js
@@ -1,17 +1,0 @@
-/**
- * @file
- * Convivial Profiler library source plugin.
- *
- * Copyright Morpht Pty Ltd 2020-2022.
- */
-
-(function (window, document) {
-  window.convivialProfiler = window.convivialProfiler || {};
-  window.convivialProfiler.profilerSource = window.convivialProfiler.profilerSource || {};
-  window.convivialProfiler.profilerSource.meta = function (profiler, source, values) {
-    var elements = document.querySelectorAll('meta[name="' + source.attribute_name + '"]');
-    for (var element of elements) {
-      values.push(element.getAttribute('content'));
-    }
-  }
-})(window, document);
